Allows to send emails with the grid action button using emails from the records.
Added options to the distribution list dropdown menu
that are names of the fields that contain emails

/!\ This is REALLY BAD, for many reasons, including but not limited to: /!\
 - it requires the field to include 'email' in its name
 - in some languages (specially languages that do not use the latin alphabet) it might not be 'email'
 - some fields might have 'email' in them but not be valid email addresses.

Added a query to the grid queries to fetch the fields and records.
Added a bit of logic to the action button, if it can't find a distribution list
with the provided name, to try to find a field with the provided name.

For it to be a bit more robust, we should probably add an 'email' tag to the fields
that contain emails, and then use that tag to find the fields.