Added a way to add an avatar to the summary card.
The avatar is added to the summary card by adding a new field to the summary card settings. The field is a tagbox field that accepts a image urls.
The URL is then used to display the avatars in the summary card.
The user can set a limit to the number of avatars that can be displayed in the summary card.

Since Angular doesn't allow us to insert HTML directly, we have to use the DomSanitizer to sanitize the HTML before inserting it into the DOM,
and make the avatar group from scratch. There is probably a better way to do this, but this is the best I could come up with 😔.

The tinyMCE editor was updated with a new button to add the avatar group to the summary card.
The user can also add the avatar group by adding the following element to the summary card settings:
{{avatar.[name_of_tagbox_field] [width] [height] [limit]}}

The tinyMCE config was updated to created all the callback functions needed to add the avatar group to the summary card.
The config file is getting quite large, so I think we should refactor it a bit to make it more readable.