AB#67717: Query parameters context dashboard
Changed the url of a dashboard to include the query parameters. This is
useful to be able to share a dashboard with a specific context.

Ex: application/appId/dashboard/baseDashboardId?record=recordId
Ex: application/appId/dashboard/baseDashboardId?element=elementId

In the noOnInit of the dashboard, we check if there is a query parameter
and if so, we load the parent dashboard, we retrieve the id of the contextual
dashboard from it and we load it.

When the user changes the context, we update the url with the new context.