Fixed an issue where the previously selected item was showing in the option list when the list was filtered by the search input, even though it should've been filtered out.
To make this work, I had to change the way the <ui-select-menu> component
was handling the optionList.onChanges event. Previously, it would set the
label (displayTriggerText) to the corresponding item in the option list.
However, when the option list is filtered, the selected item may not be in,
thus making the label empty.
If the selected value doesn't exist in the option list, the label should
stay the same as the previous one, hopefully being the label of the selected item.
When the control is emptied, the selectedValues array is emptied as well,
thus the label should be empty.