AB#67195: Fixed styling issues Copied tailwind utility classes needed for the display of avatars to the scss file of the html-widget-content component. This is not a great fix, it will very likely break somewhere in the future.
The issue is that we need tailwind to display elements the way we do in other components,
but we can't use tailwind in the html-widget-content component its purpose is to have clean styles.
(or to be more precise, to have the same styles as the TinyMCE editor, which ofc doesn't use tailwind)

One issue that might occur is that since not every web browser uses the same default styles for html elements,
the html-widget-content component might look different in different browsers.
Tailwind base styles are supposed to fix this, but since we don't use tailwind in the html-widget-content component,
we cannot use the base styles either.