fix/AB#72417_right_sidenav_wrong_height
Removed the height property on sidenav-container.component.ts (and its link to the html file)
to fix the issue of the left sidenav was too short when opened.

Since the height property was computed from the parent's element height, it was omitting the navbar height, and the tailwind css
'h-full' was overwritten by the inline height property.

This issue is fixed in the 2.0.x by using the tailwind.config.js file 'important' property, but not in the 2.1.x.