fix/AB#73095: Custom styling of widgets & application does not apply to text widget / summary card
Added SafeApplicationService to summary cards and text widgets to allow for custom styling of the widgets and application to be applied to the text widget and summary card.
The card custom styling is applied first, then the custom styling from the application if the first one is not present.

(btw I just noticed I got the name of the branch wrong)